refactor proposal report request route
