+++ src/app/router_registry.py
@@ -205,6 +205,7 @@ from app.routers.proposal_operation_support_lookups import (
     router as proposal_operation_support_lookups_router,
 )
 from app.routers.proposal_operations import router as proposal_operations_router
+from app.routers.proposal_report_requests import router as proposal_report_requests_router
 from app.routers.proposal_version_commands import router as proposal_version_commands_router
 from app.routers.proposal_versions import router as proposal_versions_router
 from app.routers.proposal_workflow import router as proposal_workflow_router
@@ -294,6 +295,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(proposal_narrative_actions_router)
     app.include_router(proposal_narratives_router)
     app.include_router(proposal_delivery_router)
+    app.include_router(proposal_report_requests_router)
     app.include_router(proposal_execution_router)
     app.include_router(proposal_execution_status_router)
     app.include_router(proposal_memos_router)
